+++ cw-orch-interchain/examples/timeout_packet.rs
@@ -60,8 +60,8 @@ fn main() -> cw_orch::anyhow::Result<()> {
     let result = interchain.await_packets("juno-1", tx_resp)?;
 
     match &result.packets[0].outcome {
-        cw_orch_interchain_core::types::IbcPacketOutcome::Timeout { .. } => {}
-        cw_orch_interchain_core::types::IbcPacketOutcome::Success { .. } => {
+        cw_orch_interchain_core::results::IbcPacketOutcome::Timeout { .. } => {}
+        cw_orch_interchain_core::results::IbcPacketOutcome::Success { .. } => {
             panic!("Expected timeout")
         }
     }
