+++ packages/interchain/interchain-daemon/examples/starship_channel.rs
@@ -2,7 +2,7 @@ use cosmwasm_std::IbcOrder;
 use cw_orch_core::environment::QuerierGetter;
 use cw_orch_daemon::queriers::Ibc;
 use cw_orch_daemon::Daemon;
-use cw_orch_interchain_core::{types::ChannelCreationResult, InterchainEnv};
+use cw_orch_interchain_core::{results::ChannelCreationResult, InterchainEnv};
 use cw_orch_interchain_daemon::ChannelCreator;
 use cw_orch_starship::Starship;
 use ibc_relayer_types::core::ics24_host::identifier::PortId;
